refactor : separate car class
