+++ src/main/java/racing/Car.java
@@ -1,0 +1,19 @@
+package racing;
+
+/**
+ * @author han
+ */
+public class Car {
+    private int step = 0;
+
+    public int forward(int step) {
+        if (isForward(step)) {
+            this.step += 1;
+        }
+        return this.step;
+    }
+
+    private boolean isForward(int step) {
+        return step > 4;
+    }
+}
