+++ src/test/java/racing/CarTest.java
@@ -3,8 +3,6 @@ package racing;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -12,21 +10,6 @@ import static org.assertj.core.api.Assertions.assertThat;
  * @author han
  */
 class CarTest {
-
-    class Car {
-        private int step = 0;
-
-        public int forward(int step) {
-            if (isForward(step)) {
-                this.step += 1;
-            }
-            return this.step;
-        }
-
-        private boolean isForward(int step) {
-            return step > 4;
-        }
-    }
 
     Car car;
